Run cron in default command rather than entrypoint

+++ start-cron.sh
@@ -1,5 +1,10 @@
 #!/bin/sh
 
+if [ ! -e /crontab ]; then
+	echo "No crontab file found!"
+	exit 1
+fi
+
 # Copy crontab file into the right place
 cp /crontab /etc/crontabs/root
